Include published date in post url for old posts
To allow redirects from Wordpress to work I need to mimic the URL as it
looks like on my Wordpress site. Therefore publish dates need to be
added to the URL.

To do this, a new `frontmatter` property has been added -
`include_date_in_url`. If this is set to `true` then the date will be
included an formatted to match `YYYY/MM/DD`.

Code that has been passing in paths have now been updated to
follow the same logic. There is a lot of code duplication, so this is
an area that could be tided up in the future.