Use `title` to generate slug name
To allow use of subfolders, renaming and reordering of folders, use the
blog post title name to generate the slug. This replaces the code that
gatsby was using by default to generate the slug via a relative path.

Moving foward, this code should also look for a backup property to
allow the title name to change without changing the slug. Without this
any changes to titles will break existing URLs from external sites.