Filter published posts when creating index pages
When a number of posts have been marked as not published gatsby was
still creating index pages that included them in the count even though
they would not be displayed. This lead to index pages with no posts.

By only including published posts in the count of total posts when
generating the index pages this error was removed.