Add RSS Feeds page
Create an RSS Feeds page so there is something for the RSS button to
link to. This page is found at `/rss`.
Change the location of the created rss feeds to be `/rss/<feed>`.
Change how `Layout` displays the titlte. Now, if a title is passed in, it
is always displayed ignoring the location of the page.
Fix some links.
Tidy up `navbar.css`.