Add navbar to blog (#23)
* Add navbar to blog

Remove bio section from homepage
Add svg social links
Add some sort of resize support to navbar
Install Osward font and use for main header + navbar

* Move navbar styling into css file

* Make the navbar fully responsive

Use what I think is the `flex-box` approach to split elements in the
navbar.
The navbar will now stetch out over the center content when on a
larger screen. On a smaller screen it will take up the whole width of
the screen.

* Add RSS Feeds page

Create an RSS Feeds page so there is something for the RSS button to
link to. This page is found at `/rss`.
Change the location of the created rss feeds to be `/rss/<feed>`.
Change how `Layout` displays the titlte. Now, if a title is passed in, it
is always displayed ignoring the location of the page.
Fix some links.
Tidy up `navbar.css`.