New logo (#24)
* New logo

Create a new SVG logo.
Use the logo as the tab image.
Add it to the navbar.
Required a lot of screwing around with CSS to get this working...

* Resize navbar based on screen size

Resize the navbar because including the logo takes up extra space. If
it is not resized, on mobile phones it takes up too much vertical space.
Making the icons and the site title smaller makes it small enough that it
can be displayed on a single line.