Remove Disqus comments from blog
Disqus was slowing down the loading of blog posts. It was making a high
amount of requests including ads.

There are alternatives to Disqus but I think it is easier to not have
them at all. I don't get that many comments anyway.

It might be possible to link the post to other social media to replace
the use of comments.