Use cloudinary to host blog post cards
Might not actually use this as the cards do not look as good as I want
them to.

Plus the free version of cloudinary is suspect to me (if my views go up
again).

But mainly because I cannot get the text to center between the title
and the tag line.

I might go through the pain of recreating all my blog cards again since
I am happy with this design and can always change it for newer posts
if I want to. This is maintainable. But, it depends on the build times
to process all the images. It will also increase the size of my repo
again...