Change profile photo file name
To fix the query which is finding my profile photo by using a regex,
and is clashing with another file name causing the image on blog
posts to show the wrong image.

Since I can't figure out how to fix it, changing the name of the photo
should suffice, so that the regex can only pick up the correct image.